[BUGFIX] Do not update session data if unchanged
To prevent session write errors after redirects (like after login)
we only try to update the session data if it really has changed.

When the data is unchanged because the timestamps do not differ
when the requests are executed very shortly after one another
the database update call returns zero changed rows which is
interpreted as an error.

When the data is unchanged we do not attempt a database update
and simply return true.